pyproject.toml
[build-system]
build-backend = "uv_build"
requires = [
    "uv-build>=0.8.22,<0.9.0",
]

[project]
name = "pre-commit-hooks"
version = "0.1.0"
description = "Some useful pre-commit hooks"
readme = "README.md"
keywords = [
    "pre-commit-hooks",
]
license = "NIST-PD"
license-files = [ "LICENSE" ]
authors = [
    { name = "William P. Krekelberg", email = "wpk@nist.gov" },
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering",
]
dependencies = [
]

optional-dependencies.all = [
    "pre-commit-hooks[fill-pre-commit-hook-deps,lastversion,sync-pre-commit-deps,sync-uv-dependency-groups]",
]

optional-dependencies.fill-pre-commit-hook-deps = [
    "dependency-groups>=1.3.1",
    "packaging>=25.0",
    "pre-commit-hooks[requirements-parser,yaml]",
]
optional-dependencies.lastversion = [
    "lastversion>=3.5.8",
]
optional-dependencies.requirements-parser = [
    "requirements-parser>=0.13.0",
]
optional-dependencies.sync-pre-commit-deps = [
    "pre-commit-hooks[requirements-parser,yaml]",
]
optional-dependencies.sync-uv-dependency-groups = [
    "packaging>=25.0",
    "tomlkit>=0.13.3",
]
optional-dependencies.yaml = [
    "ruamel-yaml>=0.18.16",
]
urls.Documentation = "https://pages.nist.gov/pre-commit-hooks/"
urls.Homepage = "https://github.com/wpk-nist-gov/pre-commit-hooks"
scripts.apply-command = "pre_commit_hooks.apply_command:main"
scripts.fill-pre-commit-deps = "pre_commit_hooks.fill_pre_commit_deps:main"
scripts.sync-pre-commit-deps = "pre_commit_hooks.sync_pre_commit_deps:main"
scripts.sync-uv-dependency-groups = "pre_commit_hooks.sync_uv_dependency_groups:main"

[dependency-groups]
dev = [
    { include-group = "mypy" },
    { include-group = "nox" },
    { include-group = "pytest-accept" },
    { include-group = "test" },
    { include-group = "typecheck" },
]
test = [
    "pytest-cov",
    "pytest-sugar",
    "pytest-xdist",
    { include-group = "optional-dependencies" },
    { include-group = "pytest" },
]
# other packages/tools
optional-dependencies = [ "pre-commit-hooks[all]" ]
cog = [ "cogapp" ]
ipython = [ "ipython" ]
mypy = [ "mypy>=1.15.0", "orjson" ]
nox = [ "nox>=2025.5.1" ]
pylint = [ "pylint" ]
pytest = [ "pytest" ]
pytest-accept = [ "pytest-accept" ]
# other environments
test-extras = [
    { include-group = "test" },
]
typecheck = [
    { include-group = "optional-dependencies" },
    { include-group = "pylint" },
    { include-group = "pytest" },
]

# * uv -------------------------------------------------------------------------

[tool.uv]
managed = true
default-groups = [ "dev" ]
fork-strategy = "requires-python"
required-version = ">=0.7.15"

[tool.uv.pip]
annotation-style = "line"
generate-hashes = true

[tool.uv.dependency-groups]
dev.requires-python = ">=3.13"
# other tools
cog.requires-python = ">=3.13"
ipython.requires-python = ">=3.13"
mypy.requires-python = ">=3.13"
nox.requires-python = ">=3.13"

[tool.uv.build-backend]
source-include = [ "tests/**" ]

# * Checks ---------------------------------------------------------------------

[tool.repo-review]
ignore = [
    "GH",     # enable for actions
    "PC111",  # use doccmd instead of blacken-docs
    "PC140",  # do this externally
    "PC170",  # pygrep
    "PC901",  # ci message
    "RTD",
    "NOX103", # I prefer using "nox.options.sessions"
    "PY004",  # no docs for this
]

[tool.check-wheel-contents]
toplevel = [ "pre_commit_hooks" ]
src_dir = [ "src" ]

[tool.pyproject-fmt]
# max_supported_python = "3.13"

# * Testing --------------------------------------------------------------------

[tool.pytest.ini_options]
addopts = [
    "-ra",
    "--strict-config",
    "--strict-markers",
    "--doctest-modules",
    "--doctest-glob=*.md",
    "--pyargs",
    "--cov-config=pyproject.toml",
]
testpaths = [
    "README.md",
    "pre_commit_hooks",
    "tests",
]
minversion = "8"
log_cli_level = "INFO"
strict_xfail = "true"
filterwarnings = [ "error" ]

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
exclude_also = [
    "if TYPE_CHECKING:",
    "@overload",
    'if __name__ == "__main__":',
]
omit = [
    "*/vendored/*",
    "*/_typing.py",
]
include = [
    "src/*",
]

[tool.coverage.paths]
source = [
    "src/",
    ".nox/**/site-packages/",
]

# * typecheck ---------------------------------------------------------------------

[tool.mypy]
files = [
    "src",
    "tests",
]
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
exclude = [
    ".eggs",
    ".tox",
    ".nox",
    ".venv",
    ".numba_cache",
]
check_untyped_defs = true
strict = true
enable_error_code = [
    "ignore-without-code",
    "redundant-expr",
    "truthy-bool",
]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [ "lastversion.*" ]

# [[tool.mypy.overrides]]
# ignore_errors = true
# module = [  ]

[tool.pyrefly]
#### configuring what to type check and where to import from
project-includes = [ "src", "tests" ]

[tool.pyright]
include = [
    "src",
    "tests",
]
exclude = [
    "**/__pycache__",
    ".tox/**",
    ".nox/**",
    ".venv/**",
    "**/.mypy_cache",
    "**/.numba_cache",
]
pythonVersion = "3.11"
typeCheckingMode = "basic"
# typeCheckingMode = "standard"
# strict = [
#     "src",
#     "tests",
#     "noxfile.py",
#     "tools/*.py",
# ]
# reportDuplicateImport = true
# reportInvalidStubStatement = true
# reportOverlappingOverload = true
# reportPropertyTypeMismatch = true
# reportUntypedClassDecorator = true
# reportUntypedFunctionDecorator = true
# reportUntypedNamedTuple = true
# reportUnusedImport = true
# Extra flags
enableTypeIgnoreComments = false
reportUnnecessaryTypeIgnoreComment = true
reportUnreachable = "warning"

[tool.pytype]
inputs = [
    "src",
    "tests",
]

# * Dependency creation --------------------------------------------------------

[tool.pyproject2conda]
user-config = "config/userconfig.toml"
template-python = "requirements/py{py}-{env}"
template = "requirements/{env}"
name = "py{py}-{env}"
python = "default"
channels = "conda-forge"
style = "yaml"
deps = "pip"

[tool.pyproject2conda.dependencies]
pytest-accept = { pip = true }

[tool.pyproject2conda.envs.dev]
groups = "dev"

[tool.pyproject2conda.envs.test]
groups = "test"
python = "all"

[tool.pyproject2conda.envs.test-extras]
groups = "test-extras"
python = "all"
skip-package = true

[tool.pyproject2conda.envs.uvx-tools]
reqs = [
    # these used solely for specking tools to be run via uvx ...
    "conda-lock>=2.5.5",
    "cruft",
    "grayskull>=2.5.3",
    "nbqa>=1.8.4",
    "pre-commit>=4.0.0",
    "pre-commit-uv",
    "prek",
    "pyrefly>=0.37.0",
    "pyright>=1.1.396",
    "basedpyright>=1.31.3",
    "scriv",
    "twine>=5.0.0",
    "ty",
]
groups = [
    "cog",
    "mypy",
    "nox",
]
skip-package = true
style = "requirements"

[tool.pyproject2conda.envs.pre-commit-additional-dependencies]
reqs = [
    "rust-just",
    "pyrefly",
]
groups = [
    "cog",
]
skip-package = true
style = "requirements"

# * Other tools ----------------------------------------------------------------

[tool.scriv]
format = "md"
md_header_level = "2"
new_fragment_template = "file: changelog.d/templates/new_fragment.md.j2"

[tool.commitizen]
use_shortcuts = true

[tool.cruft]
